Get Principal components analysis.

Given a table containing data, it performs the Principal components analysis.

Usage

getPCA(
  df,
  v,
  s = NULL,
  f = NULL,
  col_pal = NULL,
  labels_on_loading = TRUE,
  PC_to_plot = c("PC1", "PC2")
)

Arguments

df

dataframe. Containing data to plot.

v

character. The names of the columns of df containing numerical data (ideally, they should be already centered and scaled!!).

s

NULL or a character of length 1. The name of the column of df containing the sample names. Pass it only if you want sample names on the score plot.

f

NULL or character of length 1. The name of the column of df containing the factor variable. Pass it only if you want colored points and ellipses on the score plot.

col_pal

a character vector containing colors. If NULL, colors from the pals package will be used (see function build_long_vector_of_colors).

labels_on_loading

logical. If TRUE, it will report the names of features the loading plot.

PC_to_plot

character of length 2. The principal components to plots on the score and loading plots.

Value

A list with 4 objects:

df_with_scores_table: the df with additional columns containing the scores.

loadings_table: the tibble containing the loadings.

score_plot: a ggplot object with the score plot.

loading_plot: a ggplot object with the loading plot.